+++ R/bnpsd.R
@@ -15,21 +15,21 @@
 #'
 #' # define population structure
 #' # FST values for k = 2 subpopulations
-#' inbr_subpops <- c(0.1, 0.3)
+#' inbr_subpops <- c( 0.1, 0.3 )
 #' # admixture proportions from 1D geography
-#' admix_proportions <- admix_prop_1d_linear(n_ind, k_subpops, sigma = 1)
+#' admix_proportions <- admix_prop_1d_linear( n_ind, k_subpops, sigma = 1 )
 #' # also available:
 #' # - admix_prop_1d_circular
 #' # - admix_prop_indep_subpops
 #'
 #' # get pop structure parameters of the admixed individuals
 #' # the coancestry matrix
-#' coancestry <- coanc_admix(admix_proportions, inbr_subpops)
+#' coancestry <- coanc_admix( admix_proportions, inbr_subpops )
 #' # FST of admixed individuals
-#' Fst <- fst_admix(admix_proportions, inbr_subpops)
+#' Fst <- fst_admix( admix_proportions, inbr_subpops )
 #'
 #' # draw all random allele freqs and genotypes
-#' out <- draw_all_admix(admix_proportions, inbr_subpops, m_loci)
+#' out <- draw_all_admix( admix_proportions, inbr_subpops, m_loci )
 #' # genotypes
 #' X <- out$X
 #' # ancestral allele frequencies (AFs)
@@ -38,13 +38,47 @@
 #' # OR... draw each vector or matrix separately
 #' # provided for additional flexibility
 #' # ancestral AFs
-#' p_anc <- draw_p_anc(m_loci)
+#' p_anc <- draw_p_anc( m_loci )
 #' # independent subpops (intermediate) AFs
-#' p_subpops <- draw_p_subpops(p_anc, inbr_subpops)
+#' p_subpops <- draw_p_subpops( p_anc, inbr_subpops )
 #' # individual-specific AFs
-#' p_ind <- make_p_ind_admix(p_subpops, admix_proportions)
+#' p_ind <- make_p_ind_admix( p_subpops, admix_proportions )
 #' # genotypes
-#' X <- draw_genotypes_admix(p_ind)
+#' X <- draw_genotypes_admix( p_ind )
+#'
+#' ### Examples with a tree for intermediate subpopulations
+#'
+#' # tree allows for correlated subpopulations
+#' # (prev examples had independent subpopulations)
+#' 
+#' # best to start by specifying tree in Newick string format
+#' tree_str <- '(S1:0.1,(S2:0.1,S3:0.1)N1:0.1)T;'
+#' # and turn it into `phylo` object using the `ape` package
+#' library(ape)
+#' tree_subpops <- read.tree( text = tree_str )
+#' # true coancestry matrix corresponding to this tree
+#' coanc_subpops <- coanc_tree( tree_subpops )
+#' 
+#' # admixture proportions from 1D geography
+#' # (constructed again but for k=3 tree)
+#' k_subpops <- nrow( coanc_subpops )
+#' admix_proportions <- admix_prop_1d_linear( n_ind, k_subpops, sigma = 0.5 )
+#' 
+#' # get pop structure parameters of the admixed individuals
+#' # the coancestry matrix
+#' coancestry <- coanc_admix( admix_proportions, coanc_subpops )
+#' # FST of admixed individuals
+#' Fst <- fst_admix( admix_proportions, coanc_subpops )
+#'
+#' # draw all random allele freqs and genotypes, tree version
+#' out <- draw_all_admix( admix_proportions, tree_subpops = tree_subpops, m_loci = m_loci )
+#' # genotypes
+#' X <- out$X
+#' # ancestral allele frequencies (AFs)
+#' p_anc <- out$p_anc
+#'
+#' # OR... draw tree subpops (intermediate) AFs separately
+#' p_subpops_tree <- draw_p_subpops_tree( p_anc, tree_subpops )
 #' 
 #' @docType package
 #' @name bnpsd